Fixed a bug in ligature generation
After adding "IO", the number of ligatures increased.
This increase caused the glyph index to exceed the 65K limit.

In order to fix this, we now start by placing the first
generated ligature at the very beginning of the Unicode Private
Use Area at U+E000 (decimal 57344).

We blithely continue writing ligatures until we encounter our
own "reserved" area starting at U+ED90 (60816). To avoid clobbering
the area reserved for our basic abugida glyphs, we then jump
to one past U+EDF7 (60919) because that is the last slot in our
reserved space.  So then we write the very next ligature at
U+EDF8 (60920) which is free and continue until we are done.

So now we barely waste any glyph index "space" at all and it
seems we are back in business with a font that works on
multiple platforms.